Renamed helper function
