+++ helpers/set-es-stocks.js
